+++ Libplanet.Tests/Blockchain/BlockChainTest.MultiAlgorithms.cs
@@ -46,27 +46,34 @@ namespace Libplanet.Tests.Blockchain
                     new DumbAction[0]
                 );
                 HashAlgorithmType invalidAlgo = HashAlgorithmType.Of<SHA1>();
-                Block<DumbAction> invalid1 = TestUtils.MineNext(
+                Block<DumbAction> invalidBlock = TestUtils.MineNext(
                     chain.Genesis,
                     _ => invalidAlgo,
-                    miner: TestUtils.ChainPrivateKey.PublicKey
+                    miner: TestUtils.ChainPrivateKey.PublicKey,
+                    difficulty: policy.GetNextBlockDifficulty(chain)
                 ).Evaluate(TestUtils.ChainPrivateKey, chain);
-                Assert.Throws<InvalidBlockPreEvaluationHashException>(() => chain.Append(invalid1));
+                Assert.Throws<InvalidBlockHashAlgorithmTypeException>(
+                    () => chain.Append(invalidBlock));
                 HashAlgorithmType validAlgo = HashAlgorithmType.Of<MD5>();
-                Block<DumbAction> valid1 = TestUtils.MineNext(
+                Block<DumbAction> validBlock = TestUtils.MineNext(
                     chain.Genesis,
                     _ => validAlgo,
-                    miner: TestUtils.ChainPrivateKey.PublicKey
+                    miner: TestUtils.ChainPrivateKey.PublicKey,
+                    difficulty: policy.GetNextBlockDifficulty(chain)
                 ).Evaluate(TestUtils.ChainPrivateKey, chain);
-                chain.Append(valid1);
+                chain.Append(validBlock);
             }
         }
 
-        private class MultiAlgoPolicy<T> : NullPolicy<T>
+        private class MultiAlgoPolicy<T> : BlockPolicy<T>
             where T : IAction, new()
         {
-            public override HashAlgorithmType GetHashAlgorithm(long index) =>
-                index % 2 == 1 ? HashAlgorithmType.Of<MD5>() : HashAlgorithmType.Of<SHA1>();
+            public MultiAlgoPolicy()
+                : base(hashAlgorithmGetter: index => index % 2 == 1
+                    ? HashAlgorithmType.Of<MD5>()
+                    : HashAlgorithmType.Of<SHA1>())
+            {
+            }
         }
     }
 }